It appears that the get_global_id() and get_global_size() functions only return 32 bit integers, as a result they overflow for larger arrays. The workaround is to calculate the global id and global size ourselves using the local size, local id, group_id and group size.

+++ src/backend/opencl/kernel/reduce_all_long.cl
@@ -10,7 +10,7 @@
 // http://www.whatmannerofburgeristhis.com/blog/opencl-vs-cuda-gpu-memory-fences/
 
 kernel void reduce_all_kernel(global To *oData, KParam oInfo,
-                              global int* retirementCount, global To *tmp, KParam tmpInfo,
+                              global unsigned* retirementCount, global To *tmp, KParam tmpInfo,
                               const global Ti *iData, KParam iInfo,
                               uint groups_x, uint groups_y, uint repeat,
                               int change_nan, To nanval) {
@@ -99,7 +99,7 @@ kernel void reduce_all_kernel(global To *oData, KParam oInfo,
 
         // Thread 0 takes a ticket
         if (tid == 0) {
-            long ticket = atomic_inc(retirementCount);
+            unsigned ticket = atomic_inc(retirementCount);
             // If the ticket ID == number of blocks, we are the last block
             amLast = (ticket == (total_blocks - 1));
         }
